# backend/openapi.json
{
  "openapi": "3.1.0",
  "paths": {
...
    "/orders/checkout": {
      "post": {
        "operationId": "OrdersController_checkout",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateGuestOrderDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "allOf": [
                    {
                      "$ref": "#/components/schemas/ApiResponseDto"
                    },
                    {
                      "properties": {
                        "data": {
                          "$ref": "#/components/schemas/OrderResponseDto"
                        }
                      }
                    }
                  ]
                }
              }
            }
          }
        },
        "summary": "Guest checkout for storefront retail customers",
        "tags": ["orders"]
      }
    },
    "/orders/admin": {
      "post": {
        "operationId": "OrdersController_createB2bOrder",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateB2bOrderDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "allOf": [
                    {
                      "$ref": "#/components/schemas/ApiResponseDto"
                    },
                    {
                      "properties": {
                        "data": {
                          "$ref": "#/components/schemas/OrderResponseDto"
                        }
                      }
                    }
                  ]
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Create official B2B order (Admin/Sales)",
        "tags": ["orders"]
      }
    },
    "/orders": {
      "get": {
        "operationId": "OrdersController_listOrders",
        "parameters": [
          {
            "name": "page",
            "required": false,
            "in": "query",
            "schema": {
              "default": 1,
              "example": 1,
              "type": "number"
            }
          },
          {
            "name": "limit",
            "required": false,
            "in": "query",
            "schema": {
              "default": 20,
              "example": 20,
              "type": "number"
            }
          },
          {
            "name": "status",
            "required": false,
            "in": "query",
            "schema": {
              "type": "string",
              "enum": [
                "PENDING",
                "PROCESSING",
                "SHIPPED",
                "DELIVERED",
                "CANCELLED"
              ]
            }
          },
          {
            "name": "paymentStatus",
            "required": false,
            "in": "query",
            "schema": {
              "type": "string",
              "enum": [
                "PENDING",
                "DEPOSIT_PAID",
                "FULLY_PAID",
                "REFUNDED",
                "FAILED"
              ]
            }
          },
          {
            "name": "paymentMethod",
            "required": false,
            "in": "query",
            "schema": {
              "type": "string",
              "enum": ["CASH", "TRADE_CREDIT", "PAYOS", "BANK_TRANSFER"]
            }
          },
          {
            "name": "userId",
            "required": false,
            "in": "query",
            "schema": {
              "example": "019fa8bc-8f4d-7000-b366-e691f45cfb91",
              "type": "string"
            }
          },
          {
            "name": "customerPhone",
            "required": false,
            "in": "query",
            "schema": {
              "example": "0901234567",
              "type": "string"
            }
          },
          {
            "name": "search",
            "required": false,
            "in": "query",
            "schema": {
              "example": "ORD-20260904",
              "type": "string"
            }
          },
          {
            "name": "startDate",
            "required": false,
            "in": "query",
            "schema": {
              "format": "date-time",
              "example": "2026-09-01T00:00:00.000Z",
              "type": "string"
            }
          },
          {
            "name": "endDate",
            "required": false,
            "in": "query",
            "schema": {
              "format": "date-time",
              "example": "2026-09-30T23:59:59.000Z",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "allOf": [
                    {
                      "$ref": "#/components/schemas/ApiResponseDto"
                    },
                    {
                      "properties": {
                        "data": {
                          "$ref": "#/components/schemas/PaginatedOrderResponseDto"
                        }
                      }
                    }
                  ]
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "List orders with filtering and pagination",
        "tags": ["orders"]
      }
    },
    "/orders/{id}": {
      "get": {
        "operationId": "OrdersController_getOrderById",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "Order UUID",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "allOf": [
                    {
                      "$ref": "#/components/schemas/ApiResponseDto"
                    },
                    {
                      "properties": {
                        "data": {
                          "$ref": "#/components/schemas/OrderResponseDto"
                        }
                      }
                    }
                  ]
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Get detailed order by ID",
        "tags": ["orders"]
      }
    },
    "/orders/{id}/status": {
      "patch": {
        "operationId": "OrdersController_updateStatus",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "Order UUID",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateOrderStatusDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "allOf": [
                    {
                      "$ref": "#/components/schemas/ApiResponseDto"
                    },
                    {
                      "properties": {
                        "data": {
                          "$ref": "#/components/schemas/OrderResponseDto"
                        }
                      }
                    }
                  ]
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Update order status along state machine",
        "tags": ["orders"]
      }
    },
    "/orders/{id}/cancel": {
      "post": {
        "operationId": "OrdersController_cancelOrder",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "Order UUID",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "allOf": [
                    {
                      "$ref": "#/components/schemas/ApiResponseDto"
                    },
                    {
                      "properties": {
                        "data": {
                          "$ref": "#/components/schemas/OrderResponseDto"
                        }
                      }
                    }
                  ]
                }
              }
            }
          }
        },
        "summary": "Cancel order and release reserved stock",
        "tags": ["orders"]
      }
    },
...
  },
  "info": {
    "title": "Hyundai E-Commerce API",
    "description": "Hyundai E-Commerce Platform API for Commercial Vehicles, B2B Dealer Quoting, Warehouse Management, and Multi-channel Payment Processing.",
    "version": "1.0.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "bearer": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http",
        "description": "Enter your JWT token in the format: Bearer <token>"
      },
      "JWT-auth": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http",
        "description": "Enter your JWT token in the format: Bearer <token>"
      }
    },
    "schemas": {
...
      "ApiResponseDto": {
        "type": "object",
        "properties": {
          "success": {
            "type": "boolean",
            "example": true
          }
        },
        "required": ["success"]
      },
...
      "OrderItemProductSummaryDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "example": "019fa8bc-8f4d-7000-b366-e691f45cfb8f"
          },
          "nameVi": {
            "type": "string",
            "example": "Máy phát điện Diesel Hyundai DHY65KSE 60kVA 3 Pha"
          },
          "nameEn": {
            "type": "string",
            "example": "Hyundai DHY65KSE 60kVA 3 Phase Diesel Generator"
          },
          "slug": {
            "type": "string",
            "example": "may-phat-dien-diesel-hyundai-dhy65kse-60kva-3-pha"
          },
          "price": {
            "type": "string",
            "example": "245000000.00"
          },
          "images": {
            "example": ["https://res.cloudinary.com/hyundai/image.jpg"],
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "totalStockCache": {
            "type": "number",
            "example": 5
          }
        },
        "required": [
          "id",
          "nameVi",
          "slug",
          "price",
          "images",
          "totalStockCache"
        ]
      },
      "OrderItemResponseDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "example": "019fa8bc-8f4d-7000-b366-e691f45cfb91"
          },
          "orderId": {
            "type": "string",
            "example": "019fa8bc-8f4d-7000-b366-e691f45cfb92"
          },
          "productId": {
            "type": "string",
            "example": "019fa8bc-8f4d-7000-b366-e691f45cfb8f"
          },
          "productName": {
            "type": "string",
            "example": "Máy phát điện Diesel Hyundai DHY65KSE 60kVA 3 Pha"
          },
          "productSku": {
            "type": "string",
            "example": "DHY65KSE"
          },
          "quantity": {
            "type": "number",
            "example": 1
          },
          "unitPrice": {
            "type": "string",
            "example": "245000000.00"
          },
          "product": {
            "$ref": "#/components/schemas/OrderItemProductSummaryDto"
          },
          "createdAt": {
            "format": "date-time",
            "type": "string",
            "example": "2026-09-04T08:00:00.000Z"
          },
          "updatedAt": {
            "format": "date-time",
            "type": "string",
            "example": "2026-09-04T08:00:00.000Z"
          }
        },
        "required": [
          "id",
          "orderId",
          "productId",
          "productName",
          "productSku",
          "quantity",
          "unitPrice",
          "createdAt",
          "updatedAt"
        ]
      },
      "OrderUserSummaryDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "example": "019fa8bc-8f4d-7000-b366-e691f45cfb90"
          },
          "fullName": {
            "type": "string",
            "example": "Nguyễn Văn Đại Lý"
          },
          "email": {
            "type": "string",
            "example": "dealer@example.com"
          },
          "phoneNumber": {
            "type": "string",
            "example": "0911223344"
          },
          "role": {
            "type": "string",
            "example": "SALES"
          }
        },
        "required": ["id", "fullName", "email", "phoneNumber", "role"]
      },
      "OrderResponseDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "example": "019fa8bc-8f4d-7000-b366-e691f45cfb92"
          },
          "orderNumber": {
            "type": "string",
            "example": "ORD-20260904-4821"
          },
          "userId": {
            "type": "string",
            "example": "019fa8bc-8f4d-7000-b366-e691f45cfb90"
          },
          "leadId": {
            "type": "string",
            "example": "019fa8bc-8f4d-7000-b366-e691f45cfb91"
          },
          "customerName": {
            "type": "string",
            "example": "Nguyễn Văn A"
          },
          "customerPhone": {
            "type": "string",
            "example": "0901234567"
          },
          "customerEmail": {
            "type": "string",
            "example": "nguyenvana@example.com"
          },
          "companyName": {
            "type": "string",
            "example": "Công ty Cổ phần Xây Dựng Số 1"
          },
          "status": {
            "type": "string",
            "example": "PENDING",
            "enum": [
              "PENDING",
              "PROCESSING",
              "SHIPPED",
              "DELIVERED",
              "CANCELLED"
            ]
          },
          "shippingFee": {
            "type": "string",
            "example": "500000.00"
          },
          "shippingAddress": {
            "type": "string",
            "example": "Kho số 4, Cảng Tiên Sa, TP. Đà Nẵng"
          },
          "totalAmount": {
            "type": "string",
            "example": "245500000.00"
          },
          "depositAmount": {
            "type": "string",
            "example": "50000000.00"
          },
          "remainingAmount": {
            "type": "string",
            "example": "195500000.00"
          },
          "paymentMethod": {
            "type": "string",
            "example": "PAYOS",
            "enum": ["CASH", "TRADE_CREDIT", "PAYOS", "BANK_TRANSFER"]
          },
          "paymentStatus": {
            "type": "string",
            "example": "PENDING",
            "enum": [
              "PENDING",
              "DEPOSIT_PAID",
              "FULLY_PAID",
              "REFUNDED",
              "FAILED"
            ]
          },
          "approvalStatus": {
            "type": "string",
            "example": "APPROVED",
            "enum": ["PENDING", "APPROVED", "REJECTED"]
          },
          "approvedBy": {
            "type": "string",
            "example": "019fa8bc-8f4d-7000-b366-e691f45cfb9c"
          },
          "note": {
            "type": "string",
            "example": "Giao trong giờ hành chính"
          },
          "createdAt": {
            "format": "date-time",
            "type": "string",
            "example": "2026-09-04T08:00:00.000Z"
          },
          "updatedAt": {
            "format": "date-time",
            "type": "string",
            "example": "2026-09-04T08:00:00.000Z"
          },
          "items": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/OrderItemResponseDto"
            }
          },
          "user": {
            "$ref": "#/components/schemas/OrderUserSummaryDto"
          }
        },
        "required": [
          "id",
          "status",
          "shippingFee",
          "shippingAddress",
          "totalAmount",
          "paymentMethod",
          "paymentStatus",
          "approvalStatus",
          "createdAt",
          "updatedAt",
          "items"
        ]
      },
      "GuestOrderItemInputDto": {
        "type": "object",
        "properties": {
          "productId": {
            "type": "string",
            "example": "019fa8bc-8f4d-7000-b366-e691f45cfb8f",
            "description": "Product UUID to purchase"
          },
          "quantity": {
            "type": "number",
            "example": 1,
            "description": "Item quantity",
            "minimum": 1
          }
        },
        "required": ["productId", "quantity"]
      },
      "CreateGuestOrderDto": {
        "type": "object",
        "properties": {
          "customerName": {
            "type": "string",
            "example": "Nguyễn Văn A",
            "description": "Customer full name"
          },
          "customerPhone": {
            "type": "string",
            "example": "0901234567",
            "description": "Customer Vietnamese contact phone number"
          },
          "customerEmail": {
            "type": "string",
            "example": "nguyenvana@example.com",
            "description": "Customer email address for notifications"
          },
          "shippingAddress": {
            "type": "string",
            "example": "Số 123 Đường Nguyễn Trãi, Phường 2, Quận 5, TP. Hồ Chí Minh",
            "description": "Delivery destination street address"
          },
          "paymentMethod": {
            "type": "string",
            "example": "PAYOS",
            "enum": ["CASH", "TRADE_CREDIT", "PAYOS", "BANK_TRANSFER"],
            "default": "PAYOS",
            "description": "Checkout payment method"
          },
          "note": {
            "type": "string",
            "example": "Giao hàng trong giờ hành chính",
            "description": "Customer delivery notes"
          },
          "items": {
            "description": "Order line items",
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/GuestOrderItemInputDto"
            }
          }
        },
        "required": [
          "customerName",
          "customerPhone",
          "shippingAddress",
          "items"
        ]
      },
      "B2bOrderItemInputDto": {
        "type": "object",
        "properties": {
          "productId": {
            "type": "string",
            "example": "019fa8bc-8f4d-7000-b366-e691f45cfb8f",
            "description": "Product UUID to purchase"
          },
          "quantity": {
            "type": "number",
            "example": 2,
            "description": "Item quantity",
            "minimum": 1
          },
          "unitPrice": {
            "type": "object",
            "example": "120000000.00",
            "description": "Custom agreed unit price overriding catalog price"
          }
        },
        "required": ["productId", "quantity"]
      },
      "CreateB2bOrderDto": {
        "type": "object",
        "properties": {
          "userId": {
            "type": "string",
            "example": "019fa8bc-8f4d-7000-b366-e691f45cfb90",
            "description": "Linked customer/dealer user UUID if registered"
          },
          "leadId": {
            "type": "string",
            "example": "019fa8bc-8f4d-7000-b366-e691f45cfb91",
            "description": "Linked CRM Lead UUID for sales attribution & tracking"
          },
          "customerName": {
            "type": "string",
            "example": "Nguyễn Văn Đại Lý",
            "description": "Customer or representative name"
          },
          "customerPhone": {
            "type": "string",
            "example": "0911223344",
            "description": "Contact phone number"
          },
          "customerEmail": {
            "type": "string",
            "example": "dealer@example.com",
            "description": "Contact email"
          },
          "companyName": {
            "type": "string",
            "example": "Công ty Cổ phần Năng Lượng Miền Trung",
            "description": "Corporate company name"
          },
          "shippingAddress": {
            "type": "string",
            "example": "Kho số 4, Cảng Tiên Sa, TP. Đà Nẵng",
            "description": "Shipping destination address"
          },
          "paymentMethod": {
            "type": "string",
            "example": "TRADE_CREDIT",
            "enum": ["CASH", "TRADE_CREDIT", "PAYOS", "BANK_TRANSFER"],
            "default": "BANK_TRANSFER",
            "description": "B2B Payment method (TRADE_CREDIT, BANK_TRANSFER, CASH)"
          },
          "shippingFee": {
            "type": "object",
            "example": "500000.00",
            "default": 0,
            "description": "Freight / shipping fee"
          },
          "depositAmount": {
            "type": "object",
            "example": "50000000.00",
            "default": 0,
            "description": "Initial deposit amount paid"
          },
          "note": {
            "type": "string",
            "example": "Giao tại công trình kèm biên bản nghiệm thu",
            "description": "Internal order notes"
          },
          "items": {
            "description": "Order line items",
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/B2bOrderItemInputDto"
            }
          }
        },
        "required": [
          "customerName",
          "customerPhone",
          "shippingAddress",
          "items"
        ]
      },
      "PaginatedOrderResponseDto": {
        "type": "object",
        "properties": {
          "items": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/OrderResponseDto"
            }
          },
          "total": {
            "type": "number",
            "example": 10
          },
          "page": {
            "type": "number",
            "example": 1
          },
          "limit": {
            "type": "number",
            "example": 20
          }
        },
        "required": ["items", "total", "page", "limit"]
      },
      "UpdateOrderStatusDto": {
        "type": "object",
        "properties": {
          "status": {
            "type": "string",
            "example": "CONFIRMED",
            "enum": [
              "PENDING",
              "PROCESSING",
              "SHIPPED",
              "DELIVERED",
              "CANCELLED"
            ],
            "description": "New order status along the state machine"
          },
          "note": {
            "type": "string",
            "example": "Đã xác nhận thanh toán chuyển khoản và sẵn sàng đóng gói",
            "description": "Reason or operational note for status change"
          }
        },
        "required": ["status"]
      },
...
    }
  }
}